Fixed a few GCC warnings.
SVN r1672 (latestzdoom)

+++ src/sv_main.cpp
@@ -1745,7 +1745,7 @@ void SERVER_SetupNewConnection( BYTESTREAM_s *pByteStream, bool bNewPlayer )
 		}
 
 		// [RC] Kick if necessary.
-		if ( sv_maxclientsperip > 0 && ulOtherClientsFromIP >= sv_maxclientsperip )
+		if ( sv_maxclientsperip > 0 && (static_cast<LONG>(ulOtherClientsFromIP) >= sv_maxclientsperip) )
 		{
 			// Printf( "Connection from %s refused: too many connections from that IP. (sv_maxclientsperip is %d.)", NETWORK_AddressToString( AddressFrom ),  sv_maxclientsperip );
 			SERVER_ConnectionError( AddressFrom, "Too many connections from this IP.", NETWORK_ERRORCODE_TOOMANYCONNECTIONSFROMIP );
